+++ montepy/numbered_object_collection.py
@@ -254,12 +254,9 @@ class NumberedObjectCollection(ABC):
             raise ValueError(f"The number must be non-negative. {number} given.")
         conflict = False
         # only can trust cache if being updated
-        if self._problem:
-            if number in self.__num_cache:
-                conflict = True
-        else:
-            if number in self.numbers:
-                conflict = True
+        if number in self.__num_cache:
+            conflict = True
+
         if conflict:
             raise NumberConflictError(
                 f"Number {number} is already in use for the collection: {type(self).__name__} by {self[number]}"
@@ -1123,17 +1120,9 @@ class NumberedObjectCollection(ABC):
         -------
         Numbered_MCNP_Object
         """
-        try:
-            ret = self.__num_cache[i]
-            if ret.number == i:
-                return ret
-        except KeyError:
-            pass
-        for obj in self._objects:
-            self.__num_cache[obj.number] = obj
-            if obj.number == i:
-                self.__num_cache[i] = obj
-                return obj
+        ret = self.__num_cache.get(i)
+        if ret is not None and ret.number == i:
+            return ret
         return default
 
     def keys(self) -> typing.Generator[int, None, None]: